+++ .dotnet/src/Custom/Chat/ChatResponseModalities.Serialization.cs
@@ -4,32 +4,32 @@ namespace OpenAI.Chat;
 
 internal static partial class ChatContentModalitiesExtensions
 {
-    internal static IList<InternalCreateChatCompletionRequestModality> ToInternalModalities(this ChatContentModalities modalities)
+    internal static IList<InternalCreateChatCompletionRequestModality> ToInternalModalities(this ChatResponseModalities modalities)
     {
         List<InternalCreateChatCompletionRequestModality> internalModalities = [];
-        if (modalities.HasFlag(ChatContentModalities.Text))
+        if (modalities.HasFlag(ChatResponseModalities.Text))
         {
             internalModalities.Add(InternalCreateChatCompletionRequestModality.Text);
         }
-        if (modalities.HasFlag(ChatContentModalities.Audio))
+        if (modalities.HasFlag(ChatResponseModalities.Audio))
         {
             internalModalities.Add(InternalCreateChatCompletionRequestModality.Audio);
         }
         return internalModalities;
     }
 
-    internal static ChatContentModalities FromInternalModalities(IEnumerable<InternalCreateChatCompletionRequestModality> internalModalities)
+    internal static ChatResponseModalities FromInternalModalities(IEnumerable<InternalCreateChatCompletionRequestModality> internalModalities)
     {
-        ChatContentModalities result = 0;
+        ChatResponseModalities result = 0;
         foreach (InternalCreateChatCompletionRequestModality internalModality in internalModalities ?? [])
         {
             if (internalModality == InternalCreateChatCompletionRequestModality.Text)
             {
-                result |= ChatContentModalities.Text;
+                result |= ChatResponseModalities.Text;
             }
             else if (internalModality == InternalCreateChatCompletionRequestModality.Audio)
             {
-                result |= ChatContentModalities.Audio;
+                result |= ChatResponseModalities.Audio;
             }
         }
         return result;